Brought all the styles up to date for the whole app and the components.
Refactored to a routed detail view by restful id.  And reused the service
among multiple components.

Added a back button that does the same thing as the browser back button.
The browser back button works in the single page scenario.  Navigating to
the direct URL still does not work right.